+++ app/Api/Routes/Auth.php
@@ -1,0 +1,10 @@
+<?php
+
+use App\Api\Controllers\AuthController;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Route;
+
+
+Route::controller(AuthController::class)->group(function () {
+    Route::post('/authenticate', 'authenticate');
+});