+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # WIP
 - ADDED Laravel v8.83.2
 - ADDED Laravel Sail (MySql, Redis)
+- UPDATED routes location
+- ADDED Authenticate Request
+- ADDED sample UserResource